+++ pkg/scheduler/queue/user_queues.go
@@ -38,13 +38,12 @@ type querier struct {
 // This struct holds user queues for pending requests. It also keeps track of connected queriers,
 // and mapping between users and queriers.
 type queues struct {
-	userQueues   map[string]*userQueue
-	userQueuesMx sync.RWMutex
-
 	// List of all users with queues, used for iteration when searching for next queue to handle.
 	// Users removed from the middle are replaced with "". To avoid skipping users during iteration, we only shrink
 	// this list when there are ""'s at the end of it.
-	users []string
+	users      []string
+	userQueues map[string]*userQueue
+	queuesMx   sync.RWMutex
 
 	// How long to wait before removing a querier which has got disconnected
 	// but hasn't notified about a graceful shutdown.
@@ -102,8 +101,8 @@ func (q *queues) len() int {
 }
 
 func (q *queues) deleteQueue(userID string) {
-	q.userQueuesMx.Lock()
-	defer q.userQueuesMx.Unlock()
+	q.queuesMx.Lock()
+	defer q.queuesMx.Unlock()
 
 	uq := q.userQueues[userID]
 	if uq == nil {
@@ -134,8 +133,8 @@ func (q *queues) getOrAddQueue(userID string, maxQueriers int) userRequestQueue 
 		maxQueriers = 0
 	}
 
-	q.userQueuesMx.Lock()
-	defer q.userQueuesMx.Unlock()
+	q.queuesMx.Lock()
+	defer q.queuesMx.Unlock()
 
 	uq := q.userQueues[userID]
 	priorityEnabled := q.limits.QueryPriority(userID).Enabled
@@ -222,6 +221,9 @@ func (q *queues) createUserRequestQueue(userID string) userRequestQueue {
 func (q *queues) getNextQueueForQuerier(lastUserIndex int, querierID string) (userRequestQueue, string, int) {
 	uid := lastUserIndex
 
+	q.queuesMx.RLock()
+	defer q.queuesMx.RUnlock()
+
 	for iters := 0; iters < len(q.users); iters++ {
 		uid = uid + 1
 
@@ -235,9 +237,6 @@ func (q *queues) getNextQueueForQuerier(lastUserIndex int, querierID string) (us
 		if u == "" {
 			continue
 		}
-
-		q.userQueuesMx.RLock()
-		defer q.userQueuesMx.RUnlock()
 
 		uq := q.userQueues[u]